+++ tests/PdfViewerPreferencesTest.php
@@ -15,7 +15,7 @@ namespace fpdf\Tests;
 
 use fpdf\Enums\PdfDirection;
 use fpdf\Enums\PdfDuplex;
-use fpdf\Enums\PdfNonFullScreenPageMode;
+use fpdf\Enums\PdfPageMode;
 use fpdf\Enums\PdfScaling;
 use fpdf\Enums\PdfVersion;
 use fpdf\PdfViewerPreferences;
@@ -40,13 +40,13 @@ final class PdfViewerPreferencesTest extends TestCase
         self::assertTrue(PdfDuplex::NONE->isDefault());
         self::assertTrue(PdfDirection::L2R->isDefault());
         self::assertTrue(PdfScaling::APP_DEFAULT->isDefault());
-        self::assertTrue(PdfNonFullScreenPageMode::USE_NONE->isDefault());
+        self::assertTrue(PdfPageMode::USE_NONE->isDefault());
     }
 
     public function testDefaultNonFullScreenPageMode(): void
     {
-        $actual = PdfNonFullScreenPageMode::getDefault();
-        self::assertSame(PdfNonFullScreenPageMode::USE_NONE, $actual);
+        $actual = PdfPageMode::getDefault();
+        self::assertSame(PdfPageMode::USE_NONE, $actual);
     }
 
     public function testDefaultScaling(): void
@@ -69,7 +69,7 @@ final class PdfViewerPreferencesTest extends TestCase
         self::assertSame(PdfDuplex::NONE, $preferences->getDuplex());
         self::assertSame(PdfDirection::L2R, $preferences->getDirection());
         self::assertSame(PdfScaling::APP_DEFAULT, $preferences->getScaling());
-        self::assertSame(PdfNonFullScreenPageMode::USE_NONE, $preferences->getNonFullScreenPageMode());
+        self::assertSame(PdfPageMode::USE_NONE, $preferences->getNonFullScreenPageMode());
     }
 
     public function testFitWindow(): void
@@ -136,7 +136,7 @@ final class PdfViewerPreferencesTest extends TestCase
         $preferences->setDuplex(PdfDuplex::NONE);
         $preferences->setDirection(PdfDirection::L2R);
         $preferences->setScaling(PdfScaling::APP_DEFAULT);
-        $preferences->setNonFullScreenPageMode(PdfNonFullScreenPageMode::USE_NONE);
+        $preferences->setNonFullScreenPageMode(PdfPageMode::USE_NONE);
 
         $actual = $preferences->getOutput();
         self::assertSame('', $actual);